CreateProject/Index View and Database created.

+++ PF6_Team1_DotNetAssignment/Models/Package.cs
@@ -8,7 +8,7 @@ namespace PF6_Team1_DotNetAssignment.Models
     public class Package
     {
         public int PackageId { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }            // decimal 
         public string Title { get; set; }
         public string Description { get; set; }
         public string Reward { get; set; }
