+++ PF6_Team1_DotNetAssignment/Options/PackageOption.cs
@@ -9,7 +9,7 @@ namespace PF6_Team1_DotNetAssignment.Options
     public class PackageOption
     {
         public int PackageId { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }                // decimal
         public string Title { get; set; }
         public string Description { get; set; }
         public string Reward { get; set; }
